requirements.txt
streamlit
Pillow
numpy
opencv-python
streamlit-image-coordinates